lots of changes to category, switched to search bar so that it submit with rest of search
