+++ client/package.json
@@ -22,6 +22,7 @@
     "react-bootstrap-dropdown": "^0.3.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.9.0",
+    "react-select": "^5.7.0",
     "three": "^0.150.1",
     "typewriter-effect": "^2.19.0"
   },
